% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table_one.R
\name{make_table_one}
\alias{make_table_one}
\title{Make Table One}
\usage{
make_table_one(df, grouping_var, num_vars = NULL,
  binary_cat_vars = NULL, multiple_cat_vars = NULL, display = c("CP",
  "C", "P"), subgroups = NULL, mean_vars_for_subgroups = NULL,
  order_of_vars = NULL, digits = 2)
}
\arguments{
\item{df}{Dataset containing covariates}

\item{grouping_var}{Variable to group by (will be columns of table)}

\item{num_vars}{Vector of numeric variables}

\item{binary_cat_vars}{Vector of binary variables (in 0-1 format)}

\item{multiple_cat_vars}{Vector of multiple level categorical variables}

\item{display}{How to display results: `CP` = counts and proportions, `C` = counts, `P` = proportions
(defaults to `CP`)}

\item{subgroups}{Vector of subgroup variables of interest (must be factors). NOTE:  If you plan to analyze
a categorical variable by itself and then also use it for subgroup analysis, it is advised to copy
and rename the columns of the categorical variables you wish to use as subgroups (ie. rename variable
to something like `variable_sub``) so Table One can be properly sorted to reflect your specification in
`order_of_vars`.  If the subgroup does not have a unique name, it will be sorted directly below the
categorical variable.}

\item{mean_vars_for_subgroups}{Vector of numeric variables from which to calculate mean and sd (must match vector
position of corresponding subgroup variable)}

\item{order_of_vars}{A vector of all variables listed in the order of the rows you desire for your
Table One.}

\item{digits}{Number of digits to round decimals}
}
\value{
A sorted Table One including numeric, categorical and subgroup variables
}
\description{
This function allows you to make a Table One for manuscript publication in the
    Public Health Sciences
}
\examples{
\dontrun{
make_table_one(df = obpv_baseline, grouping_var = obpv_quintile, num_vars = num_vars,
binary_cat_vars = binary_cat_vars, multiple_cat_vars = multiple_cat_vars, display = 'CP',
subgroups = subgroups, mean_vars_for_subgroups = mean_vars, order_of_vars = var_order, digits = 2)
}
}
