% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_categorical.R
\name{quantify_categorical}
\alias{quantify_categorical}
\title{Quantify Categorical Variable}
\usage{
quantify_categorical(covariate, df, grouping_var, type = c("multiple",
  "binary"), display = c("CP", "C", "P"), digits = 1)
}
\arguments{
\item{covariate}{Categorical variable of interest}

\item{df}{Dataset containing covariate}

\item{grouping_var}{Variable to group by (will be columns of table)}

\item{type}{Type of categorical variable: `binary` or `multiple`}

\item{display}{How to display results: `CP` = counts and proportions, `C` = counts, `P` = proportions
(defaults to `CP`)}

\item{digits}{Number of digits to round decimals}
}
\value{
A data frame summarizing counts/proportions of categorical variable at each level of
    grouping variable
}
\description{
This function allows you to calculate counts, proportions, or a combination of both for
    binary or multiple level categorical variables. NOTE: ALL CATEGORICAL VARIABLES MUST BE CONVERTED
    TO FACTORS BEFORE USING THIS FUNCTION. For binary variables, make sure vairables are coded as 0-1
    and choose `type = 'binary'`.  This will produce output for the case where the variable equals 1.
    If both levels of the binary variable are desired, choose `type = 'multiple'` which produces
    output for all levels of the categorical variable.
}
\examples{
\dontrun{
quantify_categorical_edit(covariate = Edu_4cat, obpv_baseline, grouping_var = obpv_quintile,
type = 'multiple', display = 'CP')
}

}
