% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_counts_of_subgroup.R
\name{quantify_counts_of_subgroup}
\alias{quantify_counts_of_subgroup}
\title{Quantify Counts of Subgroup}
\usage{
quantify_counts_of_subgroup(subgroup_c, count_var, df, grouping_var,
  count_display = "CP", show_pval = TRUE, digits = 1)
}
\arguments{
\item{subgroup_c}{Subgroup variable of interest (must be factor)}

\item{count_var}{Factor variable from which to calculate counts and proportions}

\item{df}{Dataset containing covariates}

\item{grouping_var}{Variable to group by (will be columns of table)}

\item{count_display}{How to display results: `CP` = counts and proportions, `C` = counts, `P` = proportions
(defaults to `CP`)}

\item{show_pval}{Logical.  Should the p-value results be displayed?}

\item{digits}{Number of digits to round decimals}
}
\value{
A data frame summarizing counts and proportions of a binary variable by grouping and specific subgroup
    variable levels
}
\description{
This function allows you to calculate  the counts and proportions of a binary variable
    by grouping and specific subgroup variable levels (ie. counts/proportions of depressed males and
    females by levels of grouping variable, say treatment arm)
}
\examples{
\dontrun{
quantify_counts_of_subgroup(grouping_var = randAssign, subgroup_c = depress, count_var = female,
count_display = 'CP', show_pval = TRUE, df = baseline, digits = 1)
}

}
