% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_numeric.R
\name{quantify_numeric}
\alias{quantify_numeric}
\title{Quantify Numeric Variable}
\usage{
quantify_numeric(covariate, df, grouping_var, num_display = "PM",
  show_pval = TRUE, digits = 1)
}
\arguments{
\item{covariate}{Numeric variable of interest}

\item{df}{Dataset containing covariate}

\item{grouping_var}{Variable to group by (will be columns of table)}

\item{num_display}{How should results be displayed? ('PM' for mean +- sd, 'PRS' for mean (sd))}

\item{show_pval}{Logical.  Should the p-value results be displayed?}

\item{digits}{Number of digits to round decimals to}
}
\value{
A data frame summarizing mean/sd of covariate at each level of grouping variable
}
\description{
This function allows you to calculate mean and sd for a numeric variable
}
\examples{
\dontrun{
quantify_numeric(covariate = age, df = obpv_baseline, grouping_var = obpv_quintile, digits = 1)
}

}
