% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_all_categorical.R
\name{summarize_all_categorical}
\alias{summarize_all_categorical}
\title{Summarize All Categorical Variables}
\usage{
summarize_all_categorical(df, binary_cat_vars, multiple_cat_vars,
  grouping_var, display = c("CP", "C", "P"), show_pval = TRUE,
  digits = 1)
}
\arguments{
\item{df}{Dataset containing covariates of interest}

\item{binary_cat_vars}{Vector of binary variables (in 0-1 format)}

\item{multiple_cat_vars}{Vector of multiple level categorical variables}

\item{grouping_var}{Variable to group by (will be columns of table)}

\item{display}{How to display results: `CP` = counts and proportions, `C` = counts, `P` = proportions
(defaults to `CP`)}

\item{digits}{Number of digits to round decimals}
}
\value{
A data frame summarizing mean/sd of covariate at each level of grouping variable
}
\description{
This function allows you to calculate counts, proportions, or a combination of both for
    all binary or multiple level categorical variables in a dataset.  The function will take two vectors
    of categorical variables split by `binary` and `multiple` levels. NOTE: ALL CATEGORICAL VARIABLES
    MUST BE CONVERTED TO FACTORS BEFORE USING THIS FUNCTION. For binary variables, make sure vairables
    are coded as 0-1.  This function will produce output for the case where the binary variable equals 1.
    If both levels of the binary variable are desired, include the binary variable in the `multiple` level
    categorical variable vector. For `display` options: `CP` = counts and proportions, `C` = counts, and
    `P` = proportions (defaults to `CP`)
}
\examples{
\dontrun{
summarize_all_categorical(df = obpv_baseline, binary_cat_vars = binary_cat_vars,
multiple_cat_vars = multiple_cat_vars, grouping_var = obpv_quintile, display = 'CP', digits = 1)
}

}
