% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_all_numeric.R
\name{summarize_all_numeric}
\alias{summarize_all_numeric}
\title{Summarize All Numeric Variables}
\usage{
summarize_all_numeric(df, num_vars, grouping_var, show_pval = TRUE,
  digits = 1)
}
\arguments{
\item{df}{Dataset containing covariates of interest}

\item{num_vars}{Vector of numeric variables}

\item{grouping_var}{Variable to group by (will be columns of table)}

\item{digits}{Number of digits to round decimals}
}
\value{
A data frame summarizing mean/sd of covariate at each level of grouping variable
}
\description{
This function allows you to calculate mean and sd for all numeric variables in a dataset
}
\examples{
\dontrun{
quantify_numeric(covariate = age, df = obpv_baseline, grouping_var = obpv_quintile, digits = 1)
}

}
