% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_means_of_subgroup.R
\name{quantify_means_of_subgroup}
\alias{quantify_means_of_subgroup}
\title{Quantify Means of Subgroup}
\usage{
quantify_means_of_subgroup(subgroup, mean_var, df, grouping_var,
  show_pval = TRUE, digits = 1)
}
\arguments{
\item{subgroup}{Subgroup variable of interest (must be factor)}

\item{mean_var}{Numeric variable from which to calculate mean and sd}

\item{df}{Dataset containing covariates}

\item{grouping_var}{Variable to group by (will be columns of table)}

\item{show_pval}{Logical.  Should the p-value results be displayed?}

\item{digits}{Number of digits to round decimals}
}
\value{
A data frame summarizing mean and sd of a numeric variable by grouping and specific subgroup
    variable levels
}
\description{
This function allows you to calculate  mean and sd of a numeric variable by grouping
    and specific subgroup variable levels (ie. mean age of those belonging to subgroup 1 or 2 by
    levels of grouping variable)
}
\examples{
\dontrun{
quantify_means_of_subgroup(subgroup = Edu_4cat, mean_var = BMI, df = obpv_baseline,
grouping_var = obpv_quintile)
}

}
