% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_all_subgroups.R
\name{summarize_all_subgroups}
\alias{summarize_all_subgroups}
\title{Summarize All Subgroups}
\usage{
summarize_all_subgroups(df, subgroups, mean_vars, grouping_var,
  num_display = "PM", show_pval = TRUE, digits = 1)
}
\arguments{
\item{df}{Dataset containing covariates of interest}

\item{subgroups}{Vector of subgroup variables of interest (must be factors)}

\item{mean_vars}{Vector of numeric variables from which to calculate mean and sd (must match vector
position of corresponding subgroup variable)}

\item{grouping_var}{Variable to group by (will be columns of table)}

\item{num_display}{How should results be displayed? ('PM' for mean +- sd, 'PRS' for mean (sd))}

\item{show_pval}{Logical.  Should the p-value results be displayed?}

\item{digits}{Number of digits to round decimals}
}
\value{
A data frame summarizing mean and sd of a numeric variable by grouping and all subgroup
    variables of interest
}
\description{
This function allows you to calculate  mean and sd of a numeric variable by grouping
    and all subgroup variables of interest (ie. mean age of those belonging to subgroup 1 or 2 by
    levels of grouping variable, mean BMI of of those belonging to subgroup 1 or 2 by
    levels of grouping variable, etc.)
}
\examples{
\dontrun{
# First create vectors of subgroups and mean_vars, positions must match
(ie. mean BMI of levels of Edu_4cat, mean age of levels of sub_senior, etc.)

subgroups <- c('Edu_4cat', 'sub_senior', 'female', 'race_black')
mean_vars <- c('BMI', 'age', 'sbp', 'BMI')

# Then run function

summarize_all_subgroups(df = obpv_baseline, subgroups = subgroups, mean_vars = mean_vars,
grouping_var = obpv_quintile, digits = 2)

}

}
